Use create react app in todo example
