+++ .gitignore
@@ -9,3 +9,4 @@ packages/**/lib
 packages/**/build
 **/__pycache__
 *.profraw
+*.tsbuildinfo
