+++ packages/rtc-node/src/index.ts
@@ -1,6 +1,4 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-export * from './client';
-export * from './manager';
-export * from './util';
+export * from "./manager";
