+++ packages/rtc-node/src/manager.ts
@@ -5,11 +5,15 @@ import { Schema, Datastore } from "@lumino/datastore";
 
 import { CollaborationClient } from "./client";
 
-export async function createDatastore(
-  url: string,
-  id: number,
-  schemas: ReadonlyArray<Schema>
-): Promise<Datastore> {
+export async function createDatastore({
+  url,
+  id,
+  schemas,
+}: {
+  url: string;
+  id: number;
+  schemas: ReadonlyArray<Schema>;
+}): Promise<Datastore> {
   const client = new CollaborationClient({ url });
   const datastore = Datastore.create({
     id,
