+++ packages/rtc-todo-example/src/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import Todo from "./Todo";
+
+ReactDOM.render(
+  <React.StrictMode>
+    <Todo />
+  </React.StrictMode>,
+  document.getElementById("root")
+);
