+++ packages/rtc-todo-example/tsconfig.json
@@ -1,0 +1,33 @@
+{
+  "compilerOptions": {
+    "target": "es6",
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
+    "allowJs": true,
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
+    "forceConsistentCasingInFileNames": true,
+    "module": "esnext",
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "outDir": "lib",
+    "rootDir": "src",
+    "jsx": "react",
+    "noEmit": true,
+    "sourceMap": true
+  },
+  "references": [
+    {
+      "path": "../rtc-node"
+    }
+  ],
+  "include": [
+    "src"
+  ]
+}
